Add your 1.2 & 1.3 codes in the StartMenu.java file
Try to maybe edit my StartMenu.java and add your codes that you guys wrote in the switch statement wherever it mentions, Akuen and Anton. Then you need to push the updated StartMenu.java file again and then the version will be updated for everyone, and we will be basically done with EF1. Last step is to get it reviewed with Greg.